Implement Exponential Histgoram metrics (#525)
Implement Exponential Histgoram metrics

+++ Sources/OpenTelemetrySdk/Metrics/Stable/Data/Internal/ExponentialHistogramBuckets.swift
@@ -1,7 +1,7 @@
 //
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
-// 
+//
 
 import Foundation
 import OpenTelemetryApi
@@ -9,6 +9,6 @@ import OpenTelemetryApi
 public protocol ExponentialHistogramBuckets {
     var scale : Int { get }
     var offset : Int { get }
-    var bucketCounts : [Int] { get }
+    var bucketCounts : [Int64] { get }
     var totalCount : Int { get }
 }
